plugins.v2/tobypasstrackers/requirements.txt
dnspython~=2.8.0
aioquic~=1.2.0